# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pdomain-prep-for-pgdp",
    "description": "Convert a folder/zip of scanned book images into a PGDP-ready submission package.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/system/defaults": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get System Defaults",
        "operationId": "get_system_defaults",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDefaults-Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "system"
        ],
        "summary": "Put System Defaults",
        "operationId": "put_system_defaults",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemDefaults-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDefaults-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "system"
        ],
        "summary": "Reset System Defaults",
        "description": "Reset the stored SystemDefaults to spec-08 defaults. Idempotent.",
        "operationId": "reset_system_defaults",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDefaults-Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/data/system/defaults/export": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Export System Defaults",
        "description": "Return the SystemDefaults as a downloadable JSON file.",
        "operationId": "export_system_defaults",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/data/system/defaults/import": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Import System Defaults",
        "description": "Replace the stored SystemDefaults with an imported JSON document.\n\nBehaves like PUT \u2014 sharing the response shape so the frontend can\n`setDraft(response)` and the next GET will agree.",
        "operationId": "import_system_defaults",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemDefaults-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDefaults-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/data/projects/{project_id}/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search Project Pages",
        "operationId": "search_project_pages",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cdn/{key}": {
      "put": {
        "summary": "Cdn Put",
        "description": "Write ``key`` (under the data root) using bytes from the request body.\n\nRequires a valid authenticated session in all auth modes (in\n``auth_mode=none`` NoneAuth always succeeds, so the dependency is a\nno-op for single-user local installs).\n\nPath traversal (``..``, absolute paths) is rejected.\nBodies exceeding ``settings.max_cdn_upload_bytes`` are rejected with 413.",
        "operationId": "upload_cdn_asset",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineState": {
        "properties": {
          "steps": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StepState"
            },
            "type": "object",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "PipelineState"
      },
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": "default"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "proof_page_count": {
            "type": "integer",
            "title": "Proof Page Count"
          },
          "config": {
            "$ref": "#/components/schemas/ProjectConfig"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          },
          "storage_prefix": {
            "type": "string",
            "title": "Storage Prefix"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "stage_artifacts_bytes": {
            "type": "integer",
            "title": "Stage Artifacts Bytes",
            "default": 0
          },
          "source_zip_bytes": {
            "type": "integer",
            "title": "Source Zip Bytes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name",
          "created_at",
          "updated_at",
          "status",
          "page_count",
          "proof_page_count",
          "config",
          "pipeline_state",
          "storage_prefix",
          "archived",
          "stage_artifacts_bytes",
          "source_zip_bytes"
        ],
        "title": "Project"
      },
...
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "ingesting",
          "configuring",
          "processing",
          "reviewing",
          "packaging",
          "complete"
        ],
        "title": "ProjectStatus"
      },
...
      "SearchHitResponse": {
        "properties": {
          "page_id": {
            "type": "string",
            "title": "Page Id"
          },
          "idx0": {
            "type": "integer",
            "title": "Idx0"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "page_id",
          "idx0",
          "snippet",
          "score"
        ],
        "title": "SearchHitResponse"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchHitResponse"
            },
            "type": "array",
            "title": "Results"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "results",
          "total_count"
        ],
        "title": "SearchResponse"
      },
...
      "StepState": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/StepStatus",
            "default": "pending"
          },
          "pages_complete": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pages Complete"
          },
          "pages_error": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Pages Error"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "pages_complete",
          "pages_error",
          "started_at",
          "completed_at",
          "job_id"
        ],
        "title": "StepState"
      },
      "StepStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "complete",
          "error"
        ],
        "title": "StepStatus"
      },
...
      "SystemDefaults-Input": {
        "properties": {
          "text_threshold": {
            "type": "integer",
            "title": "Text Threshold",
            "default": 140
          },
          "page_h_w_ratio": {
            "type": "number",
            "title": "Page H W Ratio",
            "default": 1.65
          },
          "default_fuzzy_pct": {
            "type": "number",
            "title": "Default Fuzzy Pct",
            "default": 0.02
          },
          "default_pixel_count_columns": {
            "type": "integer",
            "title": "Default Pixel Count Columns",
            "default": 150
          },
          "default_pixel_count_rows": {
            "type": "integer",
            "title": "Default Pixel Count Rows",
            "default": 75
          },
          "ocr_engine": {
            "type": "string",
            "enum": [
              "doctr",
              "tesseract"
            ],
            "title": "Ocr Engine",
            "default": "doctr"
          },
          "ocr_model_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Model Key"
          },
          "ocr_dpi": {
            "type": "integer",
            "title": "Ocr Dpi",
            "default": 150
          },
          "ocr_bbox_edge_min_words": {
            "type": "integer",
            "title": "Ocr Bbox Edge Min Words",
            "default": 5
          },
          "layout_detector": {
            "type": "string",
            "enum": [
              "none",
              "contour",
              "pp-doclayout-plus-l"
            ],
            "title": "Layout Detector",
            "default": "pp-doclayout-plus-l"
          },
          "layout_detector_confidence": {
            "type": "number",
            "title": "Layout Detector Confidence",
            "default": 0.5
          },
          "layout_checkpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Checkpoint"
          },
          "standard_scannos": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Standard Scannos"
          },
          "hyphenation_join_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hyphenation Join List"
          }
        },
        "type": "object",
        "title": "SystemDefaults"
      },
      "SystemDefaults-Output": {
        "properties": {
          "text_threshold": {
            "type": "integer",
            "title": "Text Threshold",
            "default": 140
          },
          "page_h_w_ratio": {
            "type": "number",
            "title": "Page H W Ratio",
            "default": 1.65
          },
          "default_fuzzy_pct": {
            "type": "number",
            "title": "Default Fuzzy Pct",
            "default": 0.02
          },
          "default_pixel_count_columns": {
            "type": "integer",
            "title": "Default Pixel Count Columns",
            "default": 150
          },
          "default_pixel_count_rows": {
            "type": "integer",
            "title": "Default Pixel Count Rows",
            "default": 75
          },
          "ocr_engine": {
            "type": "string",
            "enum": [
              "doctr",
              "tesseract"
            ],
            "title": "Ocr Engine",
            "default": "doctr"
          },
          "ocr_model_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Model Key"
          },
          "ocr_dpi": {
            "type": "integer",
            "title": "Ocr Dpi",
            "default": 150
          },
          "ocr_bbox_edge_min_words": {
            "type": "integer",
            "title": "Ocr Bbox Edge Min Words",
            "default": 5
          },
          "layout_detector": {
            "type": "string",
            "enum": [
              "none",
              "contour",
              "pp-doclayout-plus-l"
            ],
            "title": "Layout Detector",
            "default": "pp-doclayout-plus-l"
          },
          "layout_detector_confidence": {
            "type": "number",
            "title": "Layout Detector Confidence",
            "default": 0.5
          },
          "layout_checkpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Checkpoint"
          },
          "standard_scannos": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Standard Scannos"
          },
          "hyphenation_join_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hyphenation Join List"
          }
        },
        "type": "object",
        "required": [
          "text_threshold",
          "page_h_w_ratio",
          "default_fuzzy_pct",
          "default_pixel_count_columns",
          "default_pixel_count_rows",
          "ocr_engine",
          "ocr_model_key",
          "ocr_dpi",
          "ocr_bbox_edge_min_words",
          "layout_detector",
          "layout_detector_confidence",
          "layout_checkpoint",
          "standard_scannos",
          "hyphenation_join_list"
        ],
        "title": "SystemDefaults"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}